Readme updates, linted project, added public methods for play/pause, playlist fixes, example page fixes
