+++ src/i18n/vuetify/en_US.ts
@@ -2,6 +2,6 @@ import { en } from 'vuetify/lib/locale'
 
 export default {
     ...en,
-    locale: "English",
-    test: "Testing Localization",
+    locale: 'English',
+    test: 'Testing Localization',
 }
