+++ tests/unit/MediaPlayer.spec.js
@@ -2,12 +2,12 @@ import { shallowMount } from '@vue/test-utils'
 import MediaPlayer from '@/components/MediaPlayer.vue'
 
 describe('HelloWorld.vue', () => {
-  it('renders props.msg when passed', () => {
-    /*const msg = 'new message'
+    it('renders props.msg when passed', () => {
+        /*const msg = 'new message'
     const wrapper = shallowMount(MediaPlayer, {
       propsData: { msg }
     })
     expect(wrapper.text()).toMatch(msg)*/
-    expect(true).toBeTruthy()
-  })
+        expect(true).toBeTruthy()
+    })
 })
