+++ vue.config.js
@@ -1,13 +1,13 @@
-const { defineConfig } = require("@vue/cli-service");
+const { defineConfig } = require('@vue/cli-service')
 
 module.exports = defineConfig({
-  transpileDependencies: ["vuetify"],
-  chainWebpack: (config) => {
-    config.module
-      .rule("i18n")
-      .resourceQuery(/blockType=i18n/)
-      .type("javascript/auto")
-      .use("i18n")
-      .loader("@intlify/vue-i18n-loader");
-  },
-});
+    transpileDependencies: ['vuetify'],
+    chainWebpack: (config) => {
+        config.module
+            .rule('i18n')
+            .resourceQuery(/blockType=i18n/)
+            .type('javascript/auto')
+            .use('i18n')
+            .loader('@intlify/vue-i18n-loader')
+    },
+})
